requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.4
et-xmlfile==1.1.0
idna==3.3
install==1.3.5
linecache2==1.0.0
lxml==4.9.1
openpyxl==3.0.10
pdfminer==20191125
pdfminer.six==20220524
pycparser==2.21
pycryptodome==3.15.0
pyodbc==4.0.32
python-dotenv==0.19.2
pywin32==303
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.26.8